Insert the path into the session when creating it (#398)
* Insert the request into the session when creating it instead of the cookies

This way, the Model not only has access to the cookies, but also the arguments,
the path, the protocol, the remote_ip etc.

This follow a similar logic of other web frameworks that, even though they
provide powerful routing mechanisms and pre compute handler arguments, still let
the underlying application get access to the full request. (see
https://docs.djangoproject.com/en/1.11/ref/request-response/, or
http://flask.pocoo.org/docs/0.12/quickstart/#accessing-request-data).

* Make request a property

So that it can have a documentation

* Initialize the cookies and the request in the Session initializer

This way, there is no need to finish the construction of the
session object after its creation.