fixes #480, eliminates extra _update_participants() calls
This fixes #480, but it's not perfect.  It eliminates the original
problem in #480.  The new, lesser problems come with the choice
between .user_done and .user_text.

.user_done is triggered whenever the widget loses focus, which happens
when the window loses focus, so if the widget has focus and then the
window loses focus, a spurious new_name message is broadcast.  This could
be supressed by tracking the name value to make sure it's new, but that
just complicates the demo and the spurious new_name message is really
really trivial.

.user_text seems like it might avoid the above, but it's fired before
the value of the widget is updated so it lags one char behind and would
complicate the demo even further to fix.

Overall this is better because it avoids the asyncio call and the
multiplying loops seen in #480.